Add styles for cards and replace overlay
The old style of making overlay required using position relative and
z-index for every element to make it appear on top of the overlay. This
method uses a box-shadow with the inset property, so the shadow is
appearing inside the box. This makes the code less redundant. Additional
styles are added for the cards and the mock data is replaced with better
ones.